Image reference refactor. (part 3)

+++ SerialPrograms/Source/CommonFramework/ImageTools/ImageFilter.cpp
@@ -27,7 +27,7 @@ void filter1_rgb32_range(
     const ConstImageRef& image,
     QImage& image0, uint32_t mins0, uint32_t maxs0, Color replace_with0, bool invert0
 ){
-    image0 = QImage(image.width(), image.height(), QImage::Format_ARGB32);
+    image0 = QImage((int)image.width(), (int)image.height(), QImage::Format_ARGB32);
 
     Kernels::filter_rgb32_range(
         image.data(), image.bytes_per_row(), image.width(), image.height(),
@@ -39,8 +39,8 @@ void filter2_rgb32_range(
     QImage& image0, uint32_t mins0, uint32_t maxs0, Color replace_with0, bool invert0,
     QImage& image1, uint32_t mins1, uint32_t maxs1, Color replace_with1, bool invert1
 ){
-    image0 = QImage(image.width(), image.height(), QImage::Format_ARGB32);
-    image1 = QImage(image.width(), image.height(), QImage::Format_ARGB32);
+    image0 = QImage((int)image.width(), (int)image.height(), QImage::Format_ARGB32);
+    image1 = QImage((int)image.width(), (int)image.height(), QImage::Format_ARGB32);
 
     Kernels::filter2_rgb32_range(
         image.data(), image.bytes_per_row(), image.width(), image.height(),
@@ -55,10 +55,10 @@ void filter4_rgb32_range(
     QImage& image2, uint32_t mins2, uint32_t maxs2, Color replace_with2, bool invert2,
     QImage& image3, uint32_t mins3, uint32_t maxs3, Color replace_with3, bool invert3
 ){
-    image0 = QImage(image.width(), image.height(), QImage::Format_ARGB32);
-    image1 = QImage(image.width(), image.height(), QImage::Format_ARGB32);
-    image2 = QImage(image.width(), image.height(), QImage::Format_ARGB32);
-    image3 = QImage(image.width(), image.height(), QImage::Format_ARGB32);
+    image0 = QImage((int)image.width(), (int)image.height(), QImage::Format_ARGB32);
+    image1 = QImage((int)image.width(), (int)image.height(), QImage::Format_ARGB32);
+    image2 = QImage((int)image.width(), (int)image.height(), QImage::Format_ARGB32);
+    image3 = QImage((int)image.width(), (int)image.height(), QImage::Format_ARGB32);
 
     Kernels::filter4_rgb32_range(
         image.data(), image.bytes_per_row(), image.width(), image.height(),
